move ConceptChildrenQueryBuilder to utils
